mistar/src/test/java/com/apollo/mistar/controller/LoginControllerTest.java
package com.apollo.mistar.controller;

import org.junit.Before;
import org.junit.Test;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;

/**
 * Copyright by Apollo Imaging Technology Pty Ltd.
 *
 * @Author: Hanmoi Daniel Choi
 * @Date : 12/01/13
 */
//@RunWith(SpringJUnit4ClassRunner.class)
//@ContextConfiguration(locations = {
//        "classpath:TEST-META-INF/spring/test-root-context.xml",
//        "classpath:TEST-META-INF/spring/test-mongo-context.xml"})
//@WebAppConfiguration("classpath:/META-INF/spring/mvc/servlet-context.xml")

public class LoginControllerTest {

    private MockMvc mockMvc;

    @Before
    public void setup(){
        mockMvc = MockMvcBuilders.standaloneSetup(new LoginController())
                                 .build();
    }

    @Test
    public void shouldGetWelcomePage() throws Exception {
        mockMvc.perform(get("/").accept(MediaType.TEXT_HTML))
                .andExpect(status().isOk())
                .andExpect(view().name("welcome"));

    }
}
